# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends by sharing photos of
    special moments, thanks to WASAPhoto! You can upload your photos
    directly from your PC, and they will be visible to everyone
    following you.
  version: 1.0.0
...
paths:
...
  /users/{uid}/photos:
    parameters:
    - $ref: "#/components/parameters/userid"

    post:
      tags: ["photos"]
      summary: uploads a new photo
      description: A new image is uploaded in the user's photos
      operationId: uploadPhoto
      requestBody:
        description: an image
        content:
          image/*:
            schema:
              description: |
                an object representing the image that the user want to
                publish
              type: object
              properties:
                newimage:
                  $ref: "#/components/schemas/binaryImage"
        required: true
      responses:
        "201":
          description: the photo is successfully uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/photo"
        "401":
          description: the user has no right to access this resource
        "500":
          description: the server has encountered an unknown error

      security:
        - bearerAuth: []

  #####################################################################
...
components:
  schemas:
...
    photo:
      description: |
        an object representing a photo published by someone: it includes an identifier for
        the photo, the username of the user who published the photo, the image (represented
        by its identifier and the path to the folder in which the image is stored), the date in which
        the photo was published, an array containing the partial profile of the users that like the photo,
        an array containing the comments published on the photo, how many likes and comments were left on the photo
      type: object
      properties:
        photoIdentifier:
          $ref: "#/components/schemas/photoId"
        owner:
          $ref: "#/components/schemas/username"
        image:
          $ref: "#/components/schemas/image"
        updateDate:
          description: the image's update date
          type: string
          format: date
          minLength: 10
          maxLength: 10
          example: "2023-12-09"
        likesOnPhoto:
          description: |
            an array containing the username of all the users
            that liked the photo
          type: array
          minItems: 0
          maxItems: 99999
          items:
            $ref: "#/components/schemas/partialProfile"
        likeCount:
          description: how many likes are left on the photo
          type: integer
          minimum: 0
          maximum: 99999
          example: 10
        commentsOnPhoto:
          description: an array containing all the comments published on the photo
          type: array
          minItems: 0
          maxItems: 99999
          items:
            $ref: "#/components/schemas/comment"
        commentCount:
          description: how many comments are published under the photo
          type: integer
          minimum: 0
          maximum: 99999
          example: 5

  ###################################################################
    comment:
      description: |
        an object representing an identifier for the comment, the partial profile
        of the user who wrote the comment, the id of the photo on which he wrote the comment,
        the text that he wrote as a comment and the date in which he wrote it
      type: object
      properties:
        commentIdentifier:
          $ref: "#/components/schemas/commentId"
        userWriter:
          $ref: "#/components/schemas/partialProfile"
        onPhoto:
          $ref: "#/components/schemas/photoId"
        saying:
          $ref: "#/components/schemas/commentOnReq"
        publishingDate:
          description: the date in which the comment was published
          type: string
          format: date
          minLength: 10
          maxLength: 10
          example: "2023-12-09"

    ###################################################################
    partialProfile:
      description: an object representing a user's personal infos (userid, username and profile picture)
      type: object
      properties:
        userid:
          $ref: "#/components/schemas/userid"
        username:
          $ref: "#/components/schemas/username"
        profilePict:
          $ref: "#/components/schemas/image"

    ###################################################################
    commentOnReq:
      description: the string that the user wants to publish as a comment
      type: string
      pattern: '^.*?$'
      minLength: 1
      maxLength: 300
      example: "Nice photo!"

    ###################################################################
...
    username:
      description: the username chosen by the user
      type: string
      example: Maria
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16

  ###################################################################
    userid:
      description: this is the user id returned by the doLogin function
      type: string
      example: "abcdef012345"
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      readOnly: true

      ###################################################################
...
    image:
      description: |
        an object representing the image identifier and the path to the folder in which the image is stored
      type: object
      properties:
        imageIdentifier:
          $ref: "#/components/schemas/imageId"
        pathToImage:
          description: the path to the folder in which the image is stored
          type: string
          pattern: '^[/].*?$'
          minLength: 1
          maxLength: 99999
          example: "/tmp/media/pippo/photos/123"

  ###################################################################
    imageId:
      description: the unique identifier for an image
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
    binaryImage:
      description: the image in binary format
      type: string
      format: binary
      minLength: 1
      maxLength: 99999
      example: "0110101001"

  ###################################################################
    photoId:
      description: the unique identifier for a photo published by someone
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
    commentId:
      description: the unique identifier for a photo published by someone
      type: integer
      minimum: 0
      example: 1234
      readOnly: true

  ###################################################################
  ###################################################################
  ###################################################################
  parameters:
    userid:
      name: uid
      in: path
      description: this is the user id returned by the doLogin function
      required: true
      schema:
        $ref: "#/components/schemas/userid"

  ###################################################################
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer

  ###################################################################